Bitmask

+++ hopper/_internal/cpp/flash.h
@@ -230,10 +230,10 @@ struct Flash_bwd_params : public Flash_fwd_params
     bool deterministic;
     index_t dq_accum_split_stride;
     
-    // Block sparsity mask (backward pass) — column-transposed [B, H, C, R] uint8
-    uint8_t const* ptr_block_mask;
-    int num_row_tiles;
-    int num_col_tiles;
+    // Block sparsity mask (backward pass) — packed int32 bitmask [B*H, N, num_mask_words]
+    uint32_t const* ptr_block_mask;
+    int num_row_tiles;   // num_mask_words = ceil(Tm/32), i.e. int32 words per (bh, n)
+    int num_col_tiles;   // N = number of KV blocks
 
     // Sorted CTA scheduling remap array for block sparsity
     int32_t const* work_remap = nullptr;
